Add: Dungeon Button to antrom.ts
