+++ scene/src/controllers/ui.controller.tsx
@@ -1,13 +1,15 @@
 import ReactEcs, { ReactEcsRenderer } from '@dcl/sdk/react-ecs'
-import { LoadingUI } from '../uis/loading'
+import { LoadingUI } from '../ui/loading/loading'
 import { NpcUtilsUi } from 'dcl-npc-toolkit'
 import Announcement from '../uis/Announcement'
 import { Color4 } from '@dcl/sdk/math'
 import * as utils from '@dcl-sdk/utils'
 import { type GameController } from './game.controller'
+import { PlayDungeonUI } from '../ui/dungeon/playDungeon'
 
 export class UIController {
   loadingUI: LoadingUI
+  playDungeonUI: PlayDungeonUI
   announcement_visible: boolean = false
   announcement: string = ''
   announcement_color: Color4 = Color4.White()
@@ -15,11 +17,13 @@ export class UIController {
   constructor(gameController: GameController) {
     this.gameController = gameController
     this.loadingUI = new LoadingUI(this)
+    this.playDungeonUI = new PlayDungeonUI(this)
     // eslint-disable-next-line @typescript-eslint/explicit-function-return-type
     const uiComponent = () => [
       this.loadingUI.mainUi(),
       NpcUtilsUi(),
-      this.announcementUI()
+      this.announcementUI(),
+      this.playDungeonUI.DungeonUI()
     ]
     ReactEcsRenderer.setUiRenderer(uiComponent)
   }
